refactor: Update Header component to include logo and responsive menu
The code changes in the Header component include adding a logo and making the menu responsive. The logo is now displayed in the header using the Link component from Next.js. The menu has been updated to be responsive, allowing it to collapse and expand on smaller screens. This improves the overall design and user experience of the application.